+++ RecordParser.Test/VariableLengthReaderBuilderTest.cs
@@ -1,7 +1,10 @@
-﻿using FluentAssertions;
+﻿using AutoFixture;
+using FluentAssertions;
 using RecordParser.Builders.Reader;
+using RecordParser.Builders.Writer;
 using RecordParser.Parsers;
 using System;
+using System.Diagnostics;
 using System.Globalization;
 using System.Linq;
 using System.Linq.Expressions;
@@ -191,6 +194,31 @@ namespace RecordParser.Test
                                             Color: Color.LightBlue));
         }
 
+        [Theory]
+        [InlineData(" ")]
+        [InlineData("2020.05.23")]
+        public void Given_record_with_nullable_struct_field_should_parse_properly(string birthday)
+        {
+            var reader = new VariableLengthReaderBuilder<(string Name, DateTime? Birthday, decimal Money, Color Color)>()
+                .Map(x => x.Name, indexColumn: 0)
+                .Map(x => x.Birthday, 1)
+                .Map(x => x.Money, 2)
+                .Map(x => x.Color, 3)
+                .Build(";");
+
+            var expectedBirthday = DateTime.TryParse(birthday, out var date)
+                                    ? date
+                                    : (DateTime?) null;
+
+            var parsed = reader.TryParse($"foo bar baz ; {birthday} ; 0123.45; LightBlue", out var result);
+
+            parsed.Should().BeTrue();
+            result.Should().BeEquivalentTo((Name: "foo bar baz",
+                                            Birthday: expectedBirthday,
+                                            Money: 123.45M,
+                                            Color: Color.LightBlue));
+        }
+
         [Fact]
         public void Given_nested_mapped_property_should_create_nested_instance_to_parse()
         {
@@ -252,32 +280,54 @@ namespace RecordParser.Test
             result.Should().BeEquivalentTo(expected);
         }
 
-        [Fact]
-        public void Parse_enum_same_way_framework()
+        [Theory]
+        // text as is
+        [InlineData("Black", Color.Black)]
+        // text uppercase
+        [InlineData("WHITE", Color.White)]
+        // text lowercase
+        [InlineData("yellow", Color.Yellow)]
+        // numeric value present in enum
+        [InlineData("3", Color.LightBlue)]
+        // numeric value NOT present in enum
+        [InlineData("777", (Color)777)]
+        public void Parse_enum_same_way_framework(string text, Color expected)
         {
             var reader = new VariableLengthReaderBuilder<Color>()
                 .Map(x => x, 0)
                 .Build(";");
 
-            // text as is
-            reader.Parse("Black").Should().Be(Color.Black);
+            reader.Parse(text).Should().Be(expected);
+        }
 
-            // text uppercase
-            reader.Parse("WHITE").Should().Be(Color.White);
+        [Theory]
+        [InlineData(FlaggedEnum.Some)]
+        [InlineData(FlaggedEnum.Another)]
+        [InlineData(FlaggedEnum.Other | FlaggedEnum.Some)]
+        [InlineData(FlaggedEnum.None | FlaggedEnum.Another)]
+        [InlineData((FlaggedEnum)777)]
+        public void Parse_flag_enum_same_way_framework(FlaggedEnum expected)
+        {
+            var reader = new VariableLengthReaderBuilder<FlaggedEnum>()
+                .Map(x => x, 0)
+                .Build(";");
 
-            // text lowercase
-            reader.Parse("yellow").Should().Be(Color.Yellow);
+            var text = expected.ToString();
 
-            // numeric value present in enum
-            reader.Parse("3").Should().Be(Color.LightBlue);
+            reader.Parse(text).Should().Be(expected);
+        }
 
-            // numeric value NOT present in enum
-            reader.Parse("777").Should().Be((Color)777);
+        [Fact]
+        public void Parse_enum_with_text_not_present_in_enum_should_be_same_way_framework()
+        {
+            var reader = new VariableLengthReaderBuilder<Color>()
+                .Map(x => x, 0)
+                .Build(";");
 
-            // text NOT present in enum
-            Action act = () => reader.Parse("foo");
+            var actualEx = AssertionExtensions.Should(() => reader.Parse("foo")).Throw<ArgumentException>().Which;
+            var expectedEx = AssertionExtensions.Should(() => Enum.Parse<Color>("foo")).Throw<ArgumentException>().Which;
 
-            act.Should().Throw<ArgumentException>().WithMessage("value foo not present in enum Color");
+            actualEx.Should().BeEquivalentTo(expectedEx, cfg => cfg.Excluding(x => x.StackTrace));
         }
 
         [Fact]
@@ -288,6 +338,53 @@ namespace RecordParser.Test
                 .Build(";");
 
             reader.Parse("777").Should().Be((EmptyEnum)777);
+        }
+
+        [Fact]
+        public void Given_variable_length_reader_used_in_multi_thread_context_parse_method_should_be_thread_safe()
+        {
+            // Arrange 
+
+            var reader = new VariableLengthReaderBuilder<(string Name, DateTime Birthday, decimal Money, Color Color)>()
+                .Map(x => x.Name, 0)
+                .Map(x => x.Birthday, 1)
+                .Map(x => x.Money, 2)
+                .Map(x => x.Color, 3)
+                .Build(" ; ");
+
+            var writer = new VariableLengthWriterBuilder<(string Name, DateTime Birthday, decimal Money, Color Color)>()
+                .Map(x => x.Name, 0)
+                .Map(x => x.Birthday, 1)
+                .Map(x => x.Money, 2)
+                .Map(x => x.Color, 3)
+                .Build(" ; ");
+
+            var lines = new Fixture()
+                .CreateMany<(string Name, DateTime Birthday, decimal Money, Color Color)>(10_000)
+                .Select(item =>
+                {
+                    Span<char> destination = stackalloc char[100];
+                    var success = writer.TryFormat(item, destination, out var charsWritten);
+                    Debug.Assert(success);
+                    var line = destination.Slice(0, charsWritten).ToString();
+                    return line;
+                })
+                .ToList();
+
+            // Act
+
+            var resultParallel = lines
+                .AsParallel()
+                .Select(line => reader.Parse(line))
+                .ToList();
+
+            var resultSequential = lines
+                .Select(line => reader.Parse(line))
+                .ToList();
+
+            // Assert
+
+            resultParallel.Should().BeEquivalentTo(resultSequential, cfg => cfg.WithStrictOrdering());
         }
     }