+++ RecordParser/Engines/Writer/WriteEngine.cs
@@ -79,7 +79,7 @@ namespace RecordParser.Engines.Writer
                     _ when prop.Type == typeof(char) || prop.Type == typeof(bool) =>
                         Expression.Call(typeof(WriteEngine), "TryFormat", Type.EmptyTypes, prop, temp, charsWritten),
 
-                    _ when prop.Type == typeof(Guid) => 
+                    _ when prop.Type == typeof(Guid) =>
                         Expression.Call(prop, "TryFormat", Type.EmptyTypes, temp, charsWritten, format),
 
                     _ => Expression.Call(prop, "TryFormat", Type.EmptyTypes, temp, charsWritten, format,
@@ -189,9 +189,7 @@ namespace RecordParser.Engines.Writer
                 })
                 .Reverse()
                 .Aggregate(Expression.Condition(
-                        test: Expression.Call(
-                               Expression.Convert(enumValue, under), "TryFormat", Type.EmptyTypes, span, charsWritten,
-                                Expression.Default(typeof(ReadOnlySpan<char>)), Expression.Constant(null, typeof(CultureInfo))),
+                        test: Expression.Call(typeof(WriteEngine), nameof(TryFormatEnumFallback), new[] { type }, enumValue, span, charsWritten),
                         ifTrue: CreateTuple(true, charsWritten),
                         ifFalse: CreateTuple(false, charsWritten)),
 
@@ -206,6 +204,23 @@ namespace RecordParser.Engines.Writer
                                 expressions: body);
 
             return blockExpr;
+        }
+
+        private static bool TryFormatEnumFallback<TEnum>(TEnum value, Span<char> destination, out int charsWritten)
+            where TEnum : struct, Enum
+        {
+            var text = value.ToString();
+
+            if (destination.Length < text.Length)
+            {
+                charsWritten = 0;
+                return false;
+            }
+
+            text.AsSpan().CopyTo(destination);
+
+            charsWritten = text.Length;
+            return true;
         }
     }
 }